Add a message form headler (MessageFormServlet.java)
